requirements.txt
django-user-accounts==2.0.3
pinax-webanalytics==3.0.0
pinax-eventlog==1.1.2
pinax-images==2.0.0

Django==1.11.5
# pinax-theme-bootstrap==8.0.1
git+https://github.com/pinax/pinax-theme-bootstrap.git@semantic
